backend/requirements.txt
# Database
sqlalchemy==2.0.0
alembic==1.13.0
asyncpg==0.30.0
psycopg2-binary>=2.9.6

# Redis
redis==5.0.1
arq==0.26.3

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings

# FastAPI and Web Framework
fastapi[all]
uvicorn
requests

# Authentication and Security
svix==1.62.0
python-jose[cryptography]
authlib
PyJWT>=2.8.0

# Testing
httpx==0.25.2
pytest==8.3.4
pytest-asyncio==0.23.8
tenacity==9.0.0
pytest_mock==3.14.0

# Utility
pydantic>=2.5.0
phonenumbers==9.0.2

# Logging
loguru>=0.5.3

# GCP storage
google-cloud-storage


# Langchain
langchain==0.3.24
langchain-openai==0.3.14
langgraph==0.3.32
langchain_community==0.3.22

sentence_transformers==4.1.0
trustcall==0.0.39

beautifulsoup4==4.13.4
tavily_python==0.6.0


pgvector==0.4.1
pypdf==5.4.0